openpgp: fix crash in TPK parser.
This fixes a crash in the TPK parser resulting from reaching and
unreachable branch. The problem is triggered by an invalid SecretSubkey
packet. The packet is parsed as Packet::Unknown(Tag::SecretSubkey,...).
The parser then calls pkt.tag() to convert the Packet into a Token.

The patch uses a match instead of tag() when converting. Closes #120.