OpenPGP User ID convention is "de facto" not "de factor"
I'm not even sure if we even need to use "de facto" when we're also
saying "convention", but i'm just doing a targeted fix here.

The fact that this one string was copied around in a dozen places
makes me a bit sad.  If there are other changes to make in this
boilerplate text, they'll also have to be made in a dozen places.

I don't know enough about how sequoia is designed to be able to
suggest a plausible boilerplate reduction strategy though.

Signed-off-by: Daniel Kahn Gillmor <dkg@fifthhorseman.net>