openpgp: Don't mark memory-safe Protected::into_vec as unsafe
In general, `unsafe` is an escape hatch for when do we suspicious but
actually memory-safe fiddling that the compiler can't understand.

Copying the secret into a raw `Vec` may be risky from the security point
of view but is not `unsafe` in the sense above. Use established practice
of using long/unwieldy names for functions that need careful thought.