ci: fail fast if codespell fails
Abort the ci run if codespell fails.

Rust-stable was marked as independent of codespell
which makes it start right away.

If codespell fails and rust-stable still runs
gitlab will show the state of the entire pipeline as pending.
In some places it also displays the state of the different stages.
Here one can see that the .pre stage already failed.
But it would be nice to get this feedback everywhere asap.

This means that the build jobs have to wait ~20sec. for codespell.
But the faster indication of syntax errors is worth that.