Include LICENSE.txt in all published crates
Having the license file in the root directory is not enough, since
cargo actions for workspace members will not consider this file.

This commit adds a symbolic link to the license file in the root
directory of all workspace members, so "cargo publish" will include
the LICENSE.txt file when publishing crates.