Add policy, rename commands, fix signatures
Now the default policy for decryption is Sequoia's standard policy. For
instance, this causes a `gpg`-encrypted message to fail on decryption,
due to the use of 3DES.

Commands have been renamed:

sign       => sign-detached
public-key => certificate

Fixed an issue with signatures (a LiteralWriter was in the writer stack
when signing, causing an empty LiteralData packet to appear in the
signature, confusing gpg). Additionally, the `S` flag was missing from
the primary key, so gpg 2 rightfully complained (but gpg 1.4 did not).

Also, the default output now is binary stdout, allowing for piping.