Use strip_{prefix,suffix} for code that's easier to follow
Instead of:

    if text.starts_with(prefix) {
        &text[prefix.len()..]
    } else {
        &text
    }

use this:

    if let Some(rest) = text.strip_prefix(prefix) {
        rest
    } else {
        &text
    }

The strip_prefix is easier to understand, and also removes the
dependency between the if condition and the computation to get the
slice after the prefix. The dependency is reasonably clear, but the
compiler does not understand it so it's plausible that a future change
might change the condition but not the slice. The approach using
strip_prefix avoids that.

This was found by the clippy lint manual_strip:
https://rust-lang.github.io/rust-clippy/master/index.html#manual_strip