Use the now-idiomatic option? syntax
This works in Rust now:

    foo?;

as a replacement for:

    if foo.is_none() {
        return None;
    }

It's similar to ? for error handling and can only be used in functions
that return an Option.

The instance of this in a macro caused the problem to be reported a
lot of time, once per time the macro was used, but luckily it can
could be fixed in only one place. Magic!

Found by clippy lint question_mark:
https://rust-lang.github.io/rust-clippy/master/index.html#question_mark