Avoid naming field setting it from variable of the same name
When creating a struct with a field foo, using a variable also named
foo, it's not necessary to name the field explicitly. Thus, instead
of:

    Self { foo: foo }

use this:

    Self { foo }

The shorter form is more idiomatic and thus less confusing to
experienced Rust programmers.

This was found by the clippy lint redundant_field_names:
https://rust-lang.github.io/rust-clippy/master/index.html#redundant_field_names

Sponsored-by: author