Allow new() without default()
It is customary in Rust to implement the Default trait for types that
can have a new method that takes no arguments. However, that's not
always wanted.

I've marked all the structures that have a new without arguments but
don't implement Default, so that if we get more of them, clippy will
warn.

Found by clippy lint new_without_default:
https://rust-lang.github.io/rust-clippy/master/index.html#new_without_default