Allow up to ten arguments to functions
By default, clippy allows up to seven arguments to a function.
Anything more and it will complain.

https://rust-lang.github.io/rust-clippy/master/index.html#too_many_arguments

Seven is already a very large number. However, changing the number of
arguments in public functions would change the API of the library, so
I've upped the limit to ten, which is enough to placate clippy.

Arguably, when we are changing the API anyway, we should consider ways
to remove the number of arguments to functions, possibly by using the
builder pattern.