Tell clippy loops that never loop are OK
We have two loops where the last statement is a return or a break,
without there being a continue. This means the loops will never loop.
This is a symptom of the code being hard to understand. We should
probably rewrite the logic without a loop, but I did not feel
confident doing that yet.

Found by clippy lint never_loop:
https://rust-lang.github.io/rust-clippy/master/index.html#never_loop