fix: change how signed file is mangled
The "binary signature" Subplot scenario was a little flaky. I could
reproduce this locally by running it repeatedly, and it would
invariably fail before the 300th repetition.

The problem is that the signature file produced by sq did not always
look like what the scenario expected. The fix is to change how the
signed file is modified: always remove the third line, after the
"BEGIN PGP SIGNATURE" and empty lines. Also, don't insist on the exit
code 1 for failure, until we've established what sq's exit codes can
be.

Fixes #786. This passes over 2000 runs of the "binary signature" test
run.