Create primary and subkeys based on configuration
Currently sq-dsm create a primary C key, an S subkey and an EtEr subkey.
Now the key creation can be controlled via the additional flag - primary-flags.
Based on the value of primary-flags:
C - existing functionality is retained, also the default if primary-flag is not specified.
CS - create a primary C+S key and an EtEr subkey.